move helper into apiutil
