+++ pkg/apiutil/apiutil.go
@@ -15,8 +15,10 @@ package apiutil
 
 import (
 	"encoding/json"
+	"fmt"
 	"io"
 	"io/ioutil"
+	"strconv"
 
 	"github.com/juju/errors"
 )
@@ -62,4 +64,23 @@ func ReadJSON(r io.ReadCloser, data interface{}) error {
 	}
 
 	return err
+}
+
+// FieldError connects an error to a particular field
+type FieldError struct {
+	error
+	field string
+}
+
+// ParseVarUint wraps strconv.ParseUint with detailed error reporting
+func ParseVarUint(vars map[string]string, varName string, base int, bitsize int) (uint64, error) {
+	str, ok := vars[varName]
+	if !ok {
+		return 0, FieldError{field: varName, error: fmt.Errorf("field %s not present", varName)}
+	}
+	parsed, err := strconv.ParseUint(str, base, bitsize)
+	if err == nil {
+		return parsed, nil
+	}
+	return parsed, FieldError{field: varName, error: err}
 }
